Add Anamorphic Flares, Water Caustics and Halftone Waves
- Anamorphic Flares (Light): drifting lights with long horizontal lens
  streaks, spectral halo rings and lens-coated ghost reflections mirrored
  through the frame centre.
- Water Caustics (Atmosphere): pool-floor light nets from light refracted
  through a moving wave surface (inverse Jacobian of the refraction map,
  pixel-size anti-aliasing), with optional dispersion and wind. Waves span
  whole cycles across the frame, so the pattern never visibly repeats.
- Halftone Waves (Graphic): analytic anti-aliased halftone dots on hex or
  square grids driven by ripples, sweeps, interference or noise.
- All three loop natively and scale with the frame size.
- Seven looks: Cinematic Flares, Golden Sun Flare, Pool Caustics, Prism
  Water, Pop Halftone, Mono Halftone Sweep, Neon Interference.
- _fxkit: bloom and colour fringing convert buffers to float32 before
  handing them to Pillow (float64 input was misread).
- Tests: 17 effects / 43 looks; the UI smoke test switches theme and
  language and opens the new looks.

+++ effects/_fxkit.py
@@ -708,7 +708,9 @@ def rgba_over(buf, layer):
 
 def _resize_channels(arr, size, resample):
     return np.stack([
-        np.asarray(Image.fromarray(np.ascontiguousarray(arr[..., c]), "F").resize(size, resample), dtype=np.float32)
+        # "F" images are float32: other dtypes would be misread as garbage.
+        np.asarray(Image.fromarray(np.ascontiguousarray(arr[..., c], dtype=np.float32), "F").resize(size, resample),
+                   dtype=np.float32)
         for c in range(arr.shape[2])
     ], axis=-1)
 
@@ -757,7 +759,7 @@ def chroma_fringe(buf, amount):
     for ch, sign in ((0, 1.0), (2, -1.0)):
         s = 1.0 + sign * px / diag
         inv = 1.0 / s
-        img = Image.fromarray(np.ascontiguousarray(buf[..., ch]), "F")
+        img = Image.fromarray(np.ascontiguousarray(buf[..., ch], dtype=np.float32), "F")
         img = img.transform((w, h), Image.AFFINE, (inv, 0.0, cx - cx * inv, 0.0, inv, cy - cy * inv), resample=BILINEAR)
         out[..., ch] = np.asarray(img, dtype=np.float32)
     return out
